add simple clang-format configuration file
Even if we don't automatically reformat the project and don't enforce
the format from the tool, it's useful to have a clang-format
configuration. It's not perfect (no tool can match our manual style),
but it's close enough to be useful for people who want to reformat part
of the code after editing.