Remove more superfluous properties on 'Disable UDP' button
The 'border_width' and 'relief' properties both cause the UI to fail to
load. The error reporting is awful; it just says:

** (gnome-control-center:194433): WARNING **: 15:29:59.108: Couldn't load builder file: (unknown)

The other layout properties only triggered warnings — in both GTK3 and
GTK4 — but clearly did nothing useful so remove those too.

And fix up the whitespace/wrapping in the tooltip text too.

Fixes: 3b0a4d552de5 ("Add support for OpenConnect's `--no-dtls` option to disable UDP")